# clean_electric_api_openapi.yaml
openapi: 3.0.4
info:
  title: Charging Infrastructure API
  description: |-
    This API enables management of users, chargers, vehicles, sessions, payments, and other core operations for 
    EV infrastructure.
    Field types and enumerations are standardized to match the database schema.
    All decimal, numeric, and bigint fields are represented as `string` in the API.
    For details, refer to the attached API and data type docs.
  termsOfService: 'https://example.com/terms/'
  contact:
    email: api-support@example.com
  license:
    name: Proprietary Internal Use
    url: 'https://example.com/license'
  version: 1.0.0
...
servers:
  - url: 'https://dev.cpo.cehive.com/'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserCompact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phoneNumber:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        role:
          type: string
...
    ChargerCompact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        model:
          type: string
        vendor:
          type: string
        serialNumber:
          type: string
        city:
          type: string
        alias:
          type: string
        maxPower:
          type: number
          format: integer
        noOfConnector:
          type: number
        tariff:
          $ref: '#/components/schemas/Tariff'
        connectors:
          type: array
          items:
            $ref: '#/components/schemas/Connector'
    Connector:
      type: object
      properties:
        id:
          type: string
          format: uuid
        chargerConnectorId:
          type: number
          format: integer
        type:
          type: string
          enum:
            - TYPE_2_AC
            - CCS2_DC
            - CHADEMO_DC
            - TYPE_1_AC
            - TYPE_6_DC
            - TYPE_7_ACDC
            - BHARAT_AC001
            - BHARAT_DC001
            - GBT_AC
            - GBT_DC
            - PANTOGRAPH_DOWN
            - PANTOGRAPH_UP
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Session:
      type: object
      properties:
        id:
          type: integer
          description: Session ID
        charger:
          $ref: '#/components/schemas/ChargerCompact'
        connector:
          $ref: '#/components/schemas/Connector'
        vehicleNo:
          type: string
          nullable: true
          description: Vehicle number
        vehicleVendor:
          type: string
          nullable: true
          description: Vehicle vendor
        vehicleModel:
          type: string
          nullable: true
          description: Vehicle model
        user:
          $ref: '#/components/schemas/UserCompact'
        startTime:
          type: string
          format: date-time
          description: Session start time
        endTime:
          type: string
          format: date-time
          nullable: true
          description: Session end time
        meterStart:
          type: string
          nullable: true
          description: Starting meter reading
        meterStop:
          type: string
          nullable: true
          description: Ending meter reading
        energyUsed:
          type: number
          nullable: true
          description: Energy consumed
        location:
          type: string
          nullable: true
          description: Location
        socStart:
          type: number
          nullable: true
          description: Starting State of Charge
        socLast:
          type: number
          nullable: true
          description: Last State of Charge
        reason:
          type: string
          nullable: true
          description: Session end reason
        status:
          type: string
          enum:
            - Preparing
            - Charging
            - SuspendedEVSE
            - SuspendedEV
            - Finishing
          description: Session status
        baseAmount:
          type: string
          nullable: true
          description: Base amount
        netCGST:
          type: string
          nullable: true
          description: CGST amount
        netSGST:
          type: string
          nullable: true
          description: SGST amount
        netIGST:
          type: string
          nullable: true
          description: IGST amount
        totalAmount:
          type: string
          nullable: true
          description: Total amount
        transaction:
          type: string
          format: uuid
          nullable: true
          description: Transaction ID
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Update timestamp
    SessionCompact:
      type: object
      properties:
        id:
          type: integer
          description: 'Session ID'
        charger:
          $ref: '#/components/schemas/ChargerCompact'
        location:
          type: string
          nullable: true
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
          nullable: true
        energyUsed:
          type: number
          format: integer
          nullable: true
        status:
          type: string
        totalAmount:
          type: string
          description: numeric as string
          nullable: true
...
    Tariff:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pricePerKWh:
          type: string
          description: numeric as string
        CGST:
          type: string
          description: numeric as string
        SGST:
          type: string
          description: numeric as string
        IGST:
          type: string
          description: numeric as string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  parameters:
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID of the user
...
    chargerId:
      name: chargerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID of the charger
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: integer
      description: ID of the session
...
  responses:
    GeneralSuccess:
      description: General success response
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
                example: Operation completed successfully
              data:
                type: object
                example: null
            required:
              - success
              - message
...
    ForbiddenError:
      description: You do not have permission to access this resource
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 'Forbidden: Access denied'
              data:
                type: object
                example: null
...
paths:
...
  /api/chargers/{chargerId}/sessions:
    get:
      tags:
        - Sessions
      summary: Get charger sessions
      description: Retrieves all sessions for a specific charger (admin/supervisor only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/chargerId'
      responses:
        '200':
          description: Sessions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionCompact'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /api/sessions/{sessionId}:
    get:
      tags:
        - Sessions
      summary: Get session by ID
      description: Retrieves a specific session by its ID
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Session'
  /api/users/{userId}/session/remote-stop-transaction:
    post:
      tags:
        - Sessions
      summary: Remote stop transaction
      description: Remotely stops a charging transaction
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chargerId:
                  type: string
                  format: uuid
                transactionId:
                  type: integer
              required:
                - chargerId
                - transactionId
      responses:
        '200':
          $ref: '#/components/responses/GeneralSuccess'
...